refactor nav drawer widget
